@@ -25,7 +25,7 @@
   },
   "devDependencies": {
     "@babel/cli": "^7.1.0",
-    "@babel/core": "^7.1.0",
+    "@babel/core": "^7.2.2",
     "@babel/plugin-proposal-class-properties": "^7.1.0",
     "@babel/plugin-proposal-export-default-from": "^7.0.0",
     "@babel/plugin-transform-runtime": "^7.1.0",
@@ -35,16 +35,16 @@
     "babel-eslint": "^10.0.1",
     "babel-jest": "^23.4.2",
     "del-cli": "^1.1.0",
-    "eslint": "^5.9.0",
+    "eslint": "^5.11.1",
     "eslint-config-mainframe": "^2.3.0",
     "flow-bin": "^0.89.0",
     "flow-copy-source": "^2.0.2",
     "flow-typed": "^2.5.1",
     "jest": "^23.5.0",
-    "jest-fetch-mock": "^2.0.1",
+    "jest-fetch-mock": "^2.1.0",
     "jest-junit": "^5.1.0",
     "jest-puppeteer": "^3.5.1",
-    "lerna": "^3.5.1",
+    "lerna": "^3.8.1",
     "prettier": "^1.15.3",
     "puppeteer": "^1.11.0"
   }
