@@ -8,9 +8,9 @@ import BaseBzz, {
   resText,
 } from '../packages/api-bzz-base'
 
-describe('bzz-base', () => {
+describe('api-bzz-base', () => {
   const TEST_URL = 'https://example.com/swarm-gateways/'
-  const bzz = new BaseBzz(TEST_URL)
+  const bzz = new BaseBzz({ url: TEST_URL })
   bzz._fetch = fetch // Injected by extending class
 
   beforeEach(() => {
@@ -91,6 +91,37 @@ describe('bzz-base', () => {
   })
 
   describe('Bzz class', () => {
+    jest.useFakeTimers()
+
+    it('_fetchTimeout() method supports timeout options', async () => {
+      const bzzTimeout = new BaseBzz({ url: TEST_URL, timeout: 5000 })
+      bzzTimeout._fetch = fetch
+
+      // Default (instance) timeout
+      fetch.mockResponseOnce('ok')
+      await bzzTimeout._fetchTimeout('test', {})
+      expect(setTimeout).toHaveBeenCalledTimes(1)
+      expect(setTimeout.mock.calls[0][1]).toBe(5000)
+
+      // Set in options
+      fetch.mockResponseOnce('ok')
+      await bzzTimeout._fetchTimeout('test', { timeout: 2000 })
+      expect(setTimeout).toHaveBeenCalledTimes(2)
+      expect(setTimeout.mock.calls[1][1]).toBe(2000)
+
+      // No timeout (set to 0)
+      fetch.mockResponseOnce('ok')
+      await bzzTimeout._fetchTimeout('test', { timeout: 0 })
+      expect(setTimeout).toHaveBeenCalledTimes(2)
+
+      // Rejects with timeout error
+      fetch.mockResponseOnce(() => new Promise(() => {})) // Never resolves
+      const req = bzzTimeout._fetchTimeout('test', {})
+      expect(setTimeout).toHaveBeenCalledTimes(3)
+      jest.runAllTimers()
+      await expect(req).rejects.toThrow('Timeout')
+    })
+
     it('getDownloadURL() creates the request URL for downloads', () => {
       // Default behavior
       expect(bzz.getDownloadURL('test', {})).toBe(`${TEST_URL}bzz:/test`)
@@ -171,11 +202,10 @@ describe('bzz-base', () => {
 
     it('_download() performs the request based on the given parameters and returns the response or error', async () => {
       fetch.mockResponseOnce('OK')
-      const res = await bzz._download(
-        'test',
-        { mode: 'raw' },
-        { accept: 'application/json' },
-      )
+      const res = await bzz._download('test', {
+        headers: { accept: 'application/json' },
+        mode: 'raw',
+      })
       expect(res.body).toBe('OK')
       const [fetchUrl, { headers }] = fetch.mock.calls[0]
       expect(fetchUrl).toBe(`${TEST_URL}bzz-raw:/test`)
@@ -201,8 +231,7 @@ describe('bzz-base', () => {
       fetch.mockResponseOnce('5678')
       const hash = await bzz._upload(
         'test',
-        { manifestHash: '1234' },
-        { 'content-length': 4 },
+        { headers: { 'content-length': 4 }, manifestHash: '1234' },
         true,
       )
       expect(hash).toBe('5678')
