@@ -1,5 +1,5 @@
-// flow-typed signature: a97477d6eee91987abadda727b6987b3
-// flow-typed version: <<STUB>>/@types/elliptic_v^6.4.0/flow_v0.89.0
+// flow-typed signature: 71c5a79d60d6146796a3d1116756c3e7
+// flow-typed version: <<STUB>>/@types/elliptic_v^6.4.1/flow_v0.89.0
 
 /**
  * This is an autogenerated libdef stub for:
