@@ -6,16 +6,22 @@ import createHex, {
   type hexInput,
   type hexValue,
 } from '@erebos/hex'
+import { interval, merge, type Observable } from 'rxjs'
+import { distinctUntilChanged, filter, flatMap } from 'rxjs/operators'
 
-import { getFeedTopic, pubKeyToAddress, signFeedUpdate } from './feed'
+import { bytesToHexValue, createFeedDigest, getFeedTopic } from './feed'
 import type {
+  BzzConfig,
   BzzMode,
   DirectoryData,
   DownloadOptions,
   FeedMetadata,
   FeedOptions,
-  KeyPair,
+  FeedParams,
+  FetchOptions,
   ListResult,
+  PollOptions,
+  SignFeedDigestFunc,
   UploadOptions,
 } from './types'
 
@@ -52,12 +58,58 @@ export const resJSON = (res: *) => resOrError(res).then(r => r.json())
 
 export const resText = (res: *) => resOrError(res).then(r => r.text())
 
+export const resSwarmHash = (res: *) => {
+  return resOrError(res)
+    .then(r => r.arrayBuffer())
+    .then(value => Buffer.from(new Uint8Array(value)).toString('hex'))
+}
+
+const defaultSignFeedDigest = () => {
+  return Promise.reject(new Error('Missing `signFeedDigest()` function'))
+}
+
 export default class BaseBzz {
+  _defaultTimeout: number
   _fetch: *
+  _signDigest: SignFeedDigestFunc
   _url: string
 
-  constructor(url: string) {
-    this._url = new URL(url).toString()
+  constructor(config: BzzConfig) {
+    const { url, timeout } = config
+    this._defaultTimeout = timeout ? timeout : 0
+    this._signDigest = config.signFeedDigest || defaultSignFeedDigest
+    this._url = url
+  }
+
+  _fetchTimeout(
+    url: string,
+    options: FetchOptions,
+    params?: Object = {},
+  ): Promise<*> {
+    const timeout =
+      options.timeout == null ? this._defaultTimeout : options.timeout
+    if (options.headers != null) {
+      params.headers = options.headers
+    }
+
+    if (timeout === 0) {
+      // No timeout
+      return this._fetch(url, params)
+    }
+
+    return new Promise((resolve, reject) => {
+      const timeoutID = setTimeout(() => {
+        reject(new Error('Timeout'))
+      }, timeout)
+      this._fetch(url, params).then(res => {
+        clearTimeout(timeoutID)
+        resolve(res)
+      })
+    })
+  }
+
+  signFeedDigest(digest: Array<number>, params?: any): Promise<hexValue> {
+    return this._signDigest(digest, params).then(bytesToHexValue)
   }
 
   getDownloadURL(
@@ -96,17 +148,17 @@ export default class BaseBzz {
 
   getFeedURL(
     userOrHash: string,
-    options?: FeedOptions = {},
+    params?: FeedParams = {},
     flag?: 'meta',
   ): string {
     let url = this._url + BZZ_MODE_PROTOCOLS.feed
-    let params = []
+    let query = []
 
     if (isHexValue(userOrHash)) {
       // user
-      params = Object.keys(options).reduce(
+      query = Object.keys(params).reduce(
         (acc, key) => {
-          const value = options[key]
+          const value = params[key]
           if (value != null) {
             acc.push(`${key}=${value}`)
           }
@@ -120,75 +172,71 @@ export default class BaseBzz {
     }
 
     if (flag != null) {
-      params.push(`${flag}=1`)
+      query.push(`${flag}=1`)
     }
 
-    return `${url}?${params.join('&')}`
+    return `${url}?${query.join('&')}`
   }
 
-  hash(domain: string, headers?: Object = {}): Promise<hexValue> {
-    return this._fetch(`${this._url}bzz-hash:/${domain}`, { headers }).then(
+  hash(domain: string, options?: FetchOptions = {}): Promise<hexValue> {
+    return this._fetchTimeout(`${this._url}bzz-hash:/${domain}`, options).then(
       resText,
     )
   }
 
-  list(
-    hash: string,
-    options?: DownloadOptions = {},
-    headers?: Object = {},
-  ): Promise<ListResult> {
+  list(hash: string, options?: DownloadOptions = {}): Promise<ListResult> {
     let url = `${this._url}bzz-list:/${hash}`
     if (options.path != null) {
       url += `/${options.path}`
     }
-    return this._fetch(url, { headers }).then(resJSON)
+    return this._fetchTimeout(url, options).then(resJSON)
   }
 
-  _download(
-    hash: string,
-    options: DownloadOptions,
-    headers?: Object = {},
-  ): Promise<*> {
+  _download(hash: string, options: DownloadOptions): Promise<*> {
     const url = this.getDownloadURL(hash, options)
-    return this._fetch(url, { headers }).then(resOrError)
+    return this._fetchTimeout(url, options).then(resOrError)
   }
 
-  download(
-    hash: string,
-    options?: DownloadOptions = {},
-    headers?: Object,
-  ): Promise<*> {
-    return this._download(hash, options, headers)
+  download(hash: string, options?: DownloadOptions = {}): Promise<*> {
+    return this._download(hash, options)
   }
 
   _upload(
     body: mixed,
     options: UploadOptions,
-    headers?: Object = {},
     raw?: boolean = false,
   ): Promise<hexValue> {
     const url = this.getUploadURL(options, raw)
-    return this._fetch(url, { body, headers, method: 'POST' }).then(resText)
+    return this._fetchTimeout(url, options, { body, method: 'POST' }).then(
+      resText,
+    )
   }
 
   uploadFile(
     data: string | Buffer,
     options?: UploadOptions = {},
-    headers?: Object = {},
   ): Promise<hexValue> {
     const body = typeof data === 'string' ? Buffer.from(data) : data
     const raw = options.contentType == null
-    headers['content-length'] = body.length
-    if (headers['content-type'] == null && !raw) {
-      headers['content-type'] = options.contentType
+
+    if (options.headers == null) {
+      options.headers = {}
     }
-    return this._upload(body, options, headers, raw)
+    options.headers['content-length'] = body.length
+    if (
+      options.headers != null &&
+      options.headers['content-type'] == null &&
+      !raw
+    ) {
+      options.headers['content-type'] = options.contentType
+    }
+
+    return this._upload(body, options, raw)
   }
 
   uploadDirectory(
     _directory: DirectoryData,
     _options?: UploadOptions,
-    _headers?: Object,
   ): Promise<hexValue> {
     return Promise.reject(new Error('Must be implemented in extending class'))
   }
@@ -196,77 +244,204 @@ export default class BaseBzz {
   upload(
     data: string | Buffer | DirectoryData,
     options?: UploadOptions = {},
-    headers?: Object = {},
   ): Promise<hexValue> {
     return typeof data === 'string' || Buffer.isBuffer(data)
       ? // $FlowFixMe: Flow doesn't understand type refinement with Buffer check
-        this.uploadFile(data, options, headers)
-      : this.uploadDirectory(data, options, headers)
+        this.uploadFile(data, options)
+      : this.uploadDirectory(data, options)
   }
 
   deleteResource(
     hash: string,
     path: string,
-    headers?: Object = {},
+    options?: FetchOptions = {},
   ): Promise<hexValue> {
     const url = this.getUploadURL({ manifestHash: hash, path })
-    return this._fetch(url, { method: 'DELETE', headers }).then(resText)
+    return this._fetchTimeout(url, options, { method: 'DELETE' }).then(resText)
   }
 
   createFeedManifest(
     user: string,
-    options?: FeedOptions = {},
-    headers?: Object = {},
+    params?: FeedParams = {},
+    options?: UploadOptions = {},
   ): Promise<hexValue> {
     const manifest = {
       entries: [
         {
           contentType: 'application/bzz-feed',
           mod_time: '0001-01-01T00:00:00Z',
-          feed: { topic: getFeedTopic(options), user },
+          feed: { topic: getFeedTopic(params), user },
         },
       ],
     }
-    return this.uploadFile(JSON.stringify(manifest), {}, headers).then(
-      hexValueType,
-    )
+    return this.uploadFile(JSON.stringify(manifest), options).then(hexValueType)
   }
 
   getFeedMetadata(
-    user: string,
-    options?: FeedOptions = {},
-    headers?: Object = {},
+    userOrHash: string,
+    params?: FeedParams = {},
+    options?: FetchOptions = {},
   ): Promise<FeedMetadata> {
-    const url = this.getFeedURL(user, options, 'meta')
-    return this._fetch(url, { headers }).then(resJSON)
+    const url = this.getFeedURL(userOrHash, params, 'meta')
+    return this._fetchTimeout(url, options).then(resJSON)
   }
 
   getFeedValue(
-    user: string,
+    userOrHash: string,
+    params?: FeedParams = {},
     options?: FeedOptions = {},
-    headers?: Object = {},
   ): Promise<*> {
-    const url = this.getFeedURL(user, options)
-    return this._fetch(url, { headers }).then(resOrError)
+    const url = this.getFeedURL(userOrHash, params)
+    return this._fetchTimeout(url, options)
+      .then(resOrError)
+      .then(res => {
+        if (options.mode === 'content-hash') {
+          return resSwarmHash(res)
+        }
+        if (options.mode === 'content-response') {
+          return resSwarmHash(res).then(hash => this.download(hash))
+        }
+        return res
+      })
+  }
+
+  pollFeedValue(
+    userOrHash: string,
+    options: PollOptions,
+    params?: FeedParams = {},
+  ): Observable<*> {
+    const sources = []
+
+    // Trigger the flow immediately by default
+    if (options.immediate !== false) {
+      sources.push([0])
+    }
+
+    // An external trigger can be provided in the options so the consumer can execute the flow when needed
+    if (options.trigger != null) {
+      sources.push(options.trigger)
+    }
+
+    const pipeline = []
+
+    // Handle whether the subscription should fail if the feed doesn't have a value
+    if (options.whenEmpty === 'error') {
+      pipeline.push(
+        flatMap(() => this.getFeedValue(userOrHash, params, options)),
+      )
+    } else {
+      const url = this.getFeedURL(userOrHash, params)
+      pipeline.push(
+        flatMap(() => {
+          return this._fetchTimeout(url, options).then(res => {
+            if (res.status === 404) {
+              return null
+            }
+            if (res.ok) {
+              return res
+            }
+            return new HTTPError(res.status, res.statusText)
+          })
+        }),
+      )
+
+      // Default behavior will emit null values, only omit them when option is set
+      if (options.whenEmpty === 'ignore') {
+        pipeline.push(filter(res => res !== null))
+      }
+    }
+
+    // In content mode, additional logic can be performed
+    if (
+      options.mode === 'content-hash' ||
+      options.mode === 'content-response'
+    ) {
+      // Parse response as Swarm content hash
+      pipeline.push(
+        flatMap(res => {
+          return res === null ? Promise.resolve(null) : resSwarmHash(res)
+        }),
+      )
+
+      // Only continue execution when the content hash has changed
+      if (options.contentChangedOnly === true) {
+        pipeline.push(distinctUntilChanged())
+      }
+
+      // Download contents from the provided hash
+      if (options.mode === 'content-response') {
+        pipeline.push(
+          flatMap(hash => {
+            return hash === null ? Promise.resolve(null) : this.download(hash)
+          }),
+        )
+      }
+    }
+
+    return merge(interval(options.interval), ...sources).pipe(...pipeline)
+  }
+
+  postSignedFeedValue(
+    user: string,
+    params: FeedParams,
+    body: Buffer,
+    options?: FetchOptions = {},
+  ): Promise<*> {
+    const url = this.getFeedURL(user, params)
+    return this._fetchTimeout(url, options, { method: 'POST', body }).then(
+      resOrError,
+    )
   }
 
   postFeedValue(
-    keyPair: KeyPair,
+    meta: FeedMetadata,
     data: hexInput,
-    options?: FeedOptions = {},
-    headers?: Object = {},
+    options?: FetchOptions,
+    signParams?: any,
   ): Promise<*> {
-    const user = pubKeyToAddress(keyPair.getPublic())
-    return this.getFeedMetadata(user, options, headers)
-      .then(meta => {
-        const body = createHex(data).toBuffer()
-        const url = this.getFeedURL(user, {
-          ...meta.feed,
-          ...meta.epoch,
-          signature: signFeedUpdate(meta, body, keyPair.getPrivate()),
-        })
-        return this._fetch(url, { method: 'POST', body, headers })
-      })
-      .then(resOrError)
+    const body = createHex(data).toBuffer()
+    const digest = createFeedDigest(meta, body)
+    return this.signFeedDigest(digest, signParams).then(signature => {
+      const params = {
+        topic: meta.feed.topic,
+        time: meta.epoch.time,
+        level: meta.epoch.level,
+        signature,
+      }
+      return this.postSignedFeedValue(meta.feed.user, params, body, options)
+    })
+  }
+
+  updateFeedValue(
+    userOrHash: string,
+    data: hexInput,
+    feedParams?: FeedParams,
+    options?: FetchOptions,
+    signParams?: any,
+  ): Promise<*> {
+    return this.getFeedMetadata(userOrHash, feedParams, options).then(meta => {
+      return this.postFeedValue(meta, data, options, signParams)
+    })
+  }
+
+  uploadFeedValue(
+    userOrHash: string,
+    data: string | Buffer | DirectoryData,
+    feedParams?: FeedParams,
+    options?: UploadOptions = {},
+    signParams?: any,
+  ): Promise<hexValue> {
+    const { contentType: _c, ...feedOptions } = options
+    return Promise.all([
+      this.upload(data, options),
+      this.getFeedMetadata(userOrHash, feedParams, feedOptions),
+    ]).then(([hash, meta]) => {
+      return this.postFeedValue(
+        meta,
+        `0x${hash}`,
+        feedOptions,
+        signParams,
+      ).then(() => hash)
+    })
   }
 }
