@@ -3,6 +3,7 @@
 import { createReadStream } from 'fs'
 import type { Readable } from 'stream'
 import BaseBzz, {
+  type BzzConfig,
   type DirectoryData,
   type FileEntry,
   type DownloadOptions,
@@ -20,29 +21,26 @@ import { isFile, writeStreamTo, extractTarStreamTo } from './fs'
 export type * from '@erebos/api-bzz-base'
 
 export default class Bzz extends BaseBzz {
-  constructor(url: string) {
-    super(url)
+  constructor(config: BzzConfig) {
+    const { url, ...cfg } = config
+    super({ ...cfg, url: new URL(url).href })
     this._fetch = fetch
   }
 
-  async _downloadTar(
-    hash: string,
-    options: DownloadOptions,
-    headers?: Object = {},
-  ): Promise<*> {
-    return await this._download(hash, options, {
-      ...headers,
-      accept: 'application/x-tar',
-    })
+  async _downloadTar(hash: string, options: DownloadOptions): Promise<*> {
+    if (options.headers == null) {
+      options.headers = {}
+    }
+    options.headers.accept = 'application/x-tar'
+    return await this._download(hash, options)
   }
 
   downloadObservable(
     hash: string,
     options?: DownloadOptions = {},
-    headers?: Object = {},
   ): Observable<FileEntry> {
     return Observable.create(observer => {
-      this._downloadTar(hash, options, headers).then(
+      this._downloadTar(hash, options).then(
         res => {
           const extract = tarStream.extract()
           extract.on('entry', (header, stream, next) => {
@@ -79,11 +77,10 @@ export default class Bzz extends BaseBzz {
   downloadDirectoryData(
     hash: string,
     options?: DownloadOptions = {},
-    headers?: Object = {},
   ): Promise<DirectoryData> {
     return new Promise((resolve, reject) => {
       const directoryData = {}
-      this.downloadObservable(hash, options, headers).subscribe({
+      this.downloadObservable(hash, options).subscribe({
         next: entry => {
           directoryData[entry.path] = { data: entry.data, size: entry.size }
         },
@@ -101,9 +98,8 @@ export default class Bzz extends BaseBzz {
     hash: string,
     toPath: string,
     options?: DownloadOptions = {},
-    headers?: Object = {},
   ): Promise<void> {
-    const res = await this._download(hash, options, headers)
+    const res = await this._download(hash, options)
     await writeStreamTo(res.body, toPath)
   }
 
@@ -111,9 +107,8 @@ export default class Bzz extends BaseBzz {
     hash: string,
     toPath: string,
     options?: DownloadOptions = {},
-    headers?: Object = {},
   ): Promise<number> {
-    const res = await this._downloadTar(hash, options, headers)
+    const res = await this._downloadTar(hash, options)
     return await extractTarStreamTo(res.body, toPath)
   }
 
@@ -121,19 +116,17 @@ export default class Bzz extends BaseBzz {
     hash: string,
     toPath: string,
     options?: DownloadOptions = {},
-    headers?: Object = {},
   ): Promise<void> {
     if (await isFile(toPath)) {
-      await this.downloadFileTo(hash, toPath, options, headers)
+      await this.downloadFileTo(hash, toPath, options)
     } else {
-      await this.downloadDirectoryTo(hash, toPath, options, headers)
+      await this.downloadDirectoryTo(hash, toPath, options)
     }
   }
 
   async uploadDirectory(
     directory: DirectoryData,
     options?: UploadOptions = {},
-    headers?: Object = {},
   ): Promise<hexValue> {
     const form = new FormData()
     Object.keys(directory).forEach(key => {
@@ -147,62 +140,69 @@ export default class Bzz extends BaseBzz {
         form.append('', file.data, { contentType: file.contentType })
       }
     }
-    return await this._upload(form, options, {
-      ...headers,
-      ...form.getHeaders(),
-    })
+    const headers = options.headers || {}
+    options.headers = { ...headers, ...form.getHeaders() }
+    return await this._upload(form, options)
+  }
+
+  async uploadFileStream(
+    stream: Readable,
+    options?: UploadOptions = {},
+  ): Promise<hexValue> {
+    const raw = options.contentType == null
+    if (!raw) {
+      if (options.headers == null) {
+        options.headers = {}
+      }
+      options.headers['content-type'] = options.contentType
+    }
+    return await this._upload(stream, options, raw)
   }
 
   async uploadFileFrom(
     path: string,
-    options?: UploadOptions = {},
-    headers?: Object = {},
+    options?: UploadOptions,
   ): Promise<hexValue> {
-    const raw = options.contentType == null
-    if (!raw) headers['content-type'] = options.contentType
-    return await this._upload(createReadStream(path), options, headers, raw)
+    return await this.uploadFileStream(createReadStream(path), options)
   }
 
   async _uploadTarStream(
     stream: Readable,
     options?: UploadOptions = {},
-    headers?: Object = {},
   ): Promise<hexValue> {
-    return await this._upload(stream, options, {
-      ...headers,
-      'content-type': 'application/x-tar',
-    })
+    if (options.headers == null) {
+      options.headers = {}
+    }
+    options.headers['content-type'] = 'application/x-tar'
+    return await this._upload(stream, options)
   }
 
   // path must be either a tar archive or a directory
   async uploadTar(
     path: string,
     options?: UploadOptions = {},
-    headers?: Object = {},
   ): Promise<hexValue> {
     const stream = (await isFile(path))
       ? createReadStream(path)
       : tarFS.pack(path)
-    return await this._uploadTarStream(stream, options, headers)
+    return await this._uploadTarStream(stream, options)
   }
 
   async uploadDirectoryFrom(
     path: string,
     options?: UploadOptions = {},
-    headers?: Object = {},
   ): Promise<hexValue> {
-    return await this._uploadTarStream(tarFS.pack(path), options, headers)
+    return await this._uploadTarStream(tarFS.pack(path), options)
   }
 
   async uploadFrom(
     path: string,
     options?: UploadOptions = {},
-    headers?: Object = {},
   ): Promise<hexValue> {
     if (await isFile(path)) {
-      return await this.uploadFileFrom(path, options, headers)
+      return await this.uploadFileFrom(path, options)
     } else {
-      return await this.uploadDirectoryFrom(path, options, headers)
+      return await this.uploadDirectoryFrom(path, options)
     }
   }
 }
