@@ -2,6 +2,7 @@
 
 import createRPC, { ipcRPC, wsRPC } from '@mainframe/rpc-node'
 import StreamRPC from '@mainframe/rpc-stream'
+import type { BzzConfig } from '@erebos/api-bzz-base'
 import BzzAPI from '@erebos/api-bzz-node'
 import PssAPI from '@erebos/api-pss'
 import BaseClient, {
@@ -12,7 +13,7 @@ import BaseClient, {
 const instantiateAPI = createInstantiateAPI(createRPC)
 
 export type SwarmConfig = ClientConfig & {
-  bzz?: string | BzzAPI,
+  bzz?: BzzConfig | BzzAPI,
   pss?: string | PssAPI,
   rpc?: StreamRPC,
 }
@@ -21,40 +22,28 @@ export default class NodeClient extends BaseClient {
   _bzz: ?BzzAPI
   _pss: ?PssAPI
 
-  constructor(config: string | SwarmConfig) {
-    if (typeof config === 'string') {
-      const rpc = createRPC(config)
-      if (rpc instanceof StreamRPC) {
-        // RPC supports stream
-        super({ rpc })
-      } else {
-        // Assume provided URL is HTTP
-        super()
-        this._bzz = new BzzAPI(config)
+  constructor(config: SwarmConfig) {
+    if (config.rpc == null) {
+      if (config.ipc != null) {
+        config.rpc = ipcRPC(config.ipc)
+      } else if (config.ws != null) {
+        config.rpc = wsRPC(config.ws)
       }
-    } else {
-      if (config.rpc == null) {
-        if (config.ipc != null) {
-          config.rpc = ipcRPC(config.ipc)
-        } else if (config.ws != null) {
-          config.rpc = wsRPC(config.ws)
-        }
-      }
-      super(config)
-
-      if (config.bzz != null) {
-        if (config.bzz instanceof BzzAPI) {
-          this._bzz = config.bzz
-        } else if (typeof config.bzz === 'string') {
-          this._bzz = new BzzAPI(config.bzz)
-        }
-      } else if (typeof config.http === 'string') {
-        this._bzz = new BzzAPI(config.http)
-      }
-
-      // $FlowFixMe: instance type
-      this._pss = instantiateAPI(config.pss, PssAPI)
     }
+    super(config)
+
+    if (config.bzz != null) {
+      if (config.bzz instanceof BzzAPI) {
+        this._bzz = config.bzz
+      } else {
+        this._bzz = new BzzAPI(config.bzz)
+      }
+    } else if (typeof config.http === 'string') {
+      this._bzz = new BzzAPI({ url: config.http })
+    }
+
+    // $FlowFixMe: instance type
+    this._pss = instantiateAPI(config.pss, PssAPI)
   }
 
   get bzz(): BzzAPI {
